requirements.txt
fastapi
uvicorn
scikit-learn
joblib
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
streamlit
requests